feat: verify url impl, need to figure out if I can get the preimage from the fedimint client when payment's completed'

+++ src/model/invoice.rs
@@ -10,20 +10,20 @@ use sqlx::FromRow;
 
 #[derive(Debug, Clone, Fields, FromRow, Serialize)]
 pub struct Invoice {
-    pub id: i32,
+    pub id: i64,
     pub bolt11: String,
     pub settled: bool,
 }
 
 #[derive(Debug, Clone, Fields, FromRow, Serialize)]
 pub struct InvoiceForCreate {
-    pub id: i32,
+    pub id: i64,
     pub bolt11: String,
 }
 
 #[derive(Debug, Clone, Fields, FromRow, Serialize)]
 pub struct InvoiceForUpdate {
-    pub id: i32,
+    pub id: i64,
     pub settled: bool,
 }
 
@@ -38,7 +38,7 @@ impl InvoiceBmc {
         base::create::<Self, _>(mm, inv_c).await
     }
 
-    pub async fn get(mm: &ModelManager, id: i32) -> Result<Invoice> {
+    pub async fn get(mm: &ModelManager, id: i64) -> Result<Invoice> {
         base::get::<Self, _>(mm, id.into()).await
     }
 
@@ -46,7 +46,7 @@ impl InvoiceBmc {
         base::update::<Self, _>(mm, inv_u.id.into(), inv_u).await
     }
 
-    pub async fn delete(mm: &ModelManager, id: i32) -> Result<()> {
+    pub async fn delete(mm: &ModelManager, id: i64) -> Result<()> {
         base::delete::<Self>(mm, id.into()).await
     }
 }
